+++ packages/typespec-rust/.scripts/tspcompile.js
@@ -35,7 +35,7 @@ const httpSpecsGroup = {
   'spector_contentneg': {input: 'payload/content-negotiation'},
   'spector_jmergepatch': {input: 'payload/json-merge-patch'},
   'spector_corepageable': {input: 'payload/pageable'},
-  //'spector_mediatype': {input: 'payload/media-type'},
+  'spector_mediatype': {input: 'payload/media-type'},
   //'spector_multipart': {input: 'payload/multipart'},
   'spector_xml': {input: 'payload/xml'},
   'spector_routes': {input: 'routes'},